backend/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==20.1.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.3
python-slugify==8.0.4
setuptools==69.2.0
text-unidecode==1.3
Werkzeug==3.0.2
wheel==0.43.0
zipp==3.18.1
